pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.7",
    "faiss-cpu>=1.13.1",
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "langchain-groq>=0.1.6",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.1",
    "openai>=0.27.10",
    "pandas>=2.3.3",
    "pydantic>=2.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
